# schemas/web-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JeRyu Web Forge API",
    "description": "BFF for the JeRyu Web Forge. Versioned under /api/v1/. See WEB_WORK_CLAUDE.md §35.7 for the canonical route map.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0-alpha"
  },
  "paths": {
...
    "/api/v1/repos": {
      "get": {
        "tags": [
          "repos"
        ],
        "operationId": "list_repos",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "host",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "family",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "repos"
        ],
        "operationId": "create_repo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepositoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created repository",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositorySummary"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/v1/repos/preview": {
      "post": {
        "tags": [
          "repos"
        ],
        "operationId": "create_repo_preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepositoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Create preview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRepositoryPreview"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
    "/api/v1/repos/{repo_id}": {
      "get": {
        "tags": [
          "repos"
        ],
        "operationId": "get_repo",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Repository detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Repository not found"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "repos"
        ],
        "operationId": "patch_repo",
        "parameters": [
          {
            "name": "repo_id",
            "in": "path",
            "description": "Stable opaque repo ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Patched repository settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositorySettings"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Repository not found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "csrf": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessSettings": {
        "type": "object",
        "required": [
          "collaborators_count",
          "teams_count",
          "deploy_keys_count",
          "app_installations_count"
        ],
        "properties": {
          "app_installations_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "collaborators_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "deploy_keys_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "teams_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "ActionRefSurrogate": {
        "type": "object",
        "required": [
          "action_id",
          "label"
        ],
        "properties": {
          "action_id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "risk": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "AgentSettings": {
        "type": "object",
        "required": [
          "autonomous_coding_enabled",
          "max_concurrent_sessions",
          "require_human_approval_for_writes",
          "allowed_agents",
          "allowed_tools",
          "evidence_required"
        ],
        "properties": {
          "allowed_agents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowed_tools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "autonomous_coding_enabled": {
            "type": "boolean"
          },
          "budget_daily_usd": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "evidence_required": {
            "type": "boolean"
          },
          "max_concurrent_sessions": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "require_human_approval_for_writes": {
            "type": "boolean"
          }
        }
      },
...
      "BranchProtectionRule": {
        "type": "object",
        "required": [
          "pattern",
          "required_checks",
          "required_approvals",
          "require_signed_commits",
          "require_conversation_resolution",
          "allow_force_pushes",
          "allow_deletions",
          "bypass_actors"
        ],
        "properties": {
          "allow_deletions": {
            "type": "boolean"
          },
          "allow_force_pushes": {
            "type": "boolean"
          },
          "bypass_actors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pattern": {
            "type": "string"
          },
          "require_conversation_resolution": {
            "type": "boolean"
          },
          "require_signed_commits": {
            "type": "boolean"
          },
          "required_approvals": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "required_checks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "CiSettings": {
        "type": "object",
        "required": [
          "artifact_retention_days",
          "log_retention_days",
          "cache_retention_days",
          "vti_enabled"
        ],
        "properties": {
          "artifact_retention_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "cache_retention_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "concurrency_limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "default_runner_pool": {
            "type": [
              "string",
              "null"
            ]
          },
          "log_retention_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "vti_enabled": {
            "type": "boolean"
          }
        }
      },
...
      "CreateRepositoryPreview": {
        "type": "object",
        "required": [
          "normalized_name",
          "target_owner",
          "visibility",
          "initial_files",
          "settings_to_apply",
          "side_effects",
          "warnings"
        ],
        "properties": {
          "initial_files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "normalized_name": {
            "type": "string"
          },
          "settings_to_apply": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "side_effects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "target_owner": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/RepositoryVisibility"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateRepositoryRequest": {
        "type": "object",
        "required": [
          "host",
          "owner",
          "name",
          "visibility",
          "initialize_readme",
          "topics",
          "dry_run"
        ],
        "properties": {
          "default_branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "dry_run": {
            "type": "boolean"
          },
          "family": {
            "type": [
              "string",
              "null"
            ]
          },
          "gitignore_template": {
            "type": [
              "string",
              "null"
            ]
          },
          "host": {
            "type": "string"
          },
          "initialize_readme": {
            "type": "boolean"
          },
          "license_template": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "template": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RepositoryId"
              }
            ]
          },
          "topics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "$ref": "#/components/schemas/RepositoryVisibility"
          }
        }
      },
...
      "EntityRefSurrogate": {
        "type": "object",
        "required": [
          "kind",
          "id"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          }
        }
      },
...
      "FeatureSettings": {
        "type": "object",
        "required": [
          "issues",
          "merge_requests",
          "wiki",
          "discussions",
          "projects",
          "packages",
          "releases",
          "ci",
          "security_advisories",
          "pages"
        ],
        "properties": {
          "ci": {
            "type": "boolean"
          },
          "discussions": {
            "type": "boolean"
          },
          "issues": {
            "type": "boolean"
          },
          "merge_requests": {
            "type": "boolean"
          },
          "packages": {
            "type": "boolean"
          },
          "pages": {
            "type": "boolean"
          },
          "projects": {
            "type": "boolean"
          },
          "releases": {
            "type": "boolean"
          },
          "security_advisories": {
            "type": "boolean"
          },
          "wiki": {
            "type": "boolean"
          }
        }
      },
...
      "GeneralSettings": {
        "type": "object",
        "required": [
          "name",
          "visibility",
          "default_branch",
          "topics",
          "archived"
        ],
        "properties": {
          "archived": {
            "type": "boolean"
          },
          "default_branch": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "homepage": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "topics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "$ref": "#/components/schemas/RepositoryVisibility"
          }
        }
      },
...
      "MergeSettings": {
        "type": "object",
        "required": [
          "allow_merge_commit",
          "allow_squash_merge",
          "allow_rebase_merge",
          "allow_auto_merge",
          "delete_branch_on_merge",
          "require_linear_history",
          "required_approvals",
          "dismiss_stale_approvals",
          "require_codeowners",
          "require_exact_sha_approval",
          "require_jeryu_merge_passport"
        ],
        "properties": {
          "allow_auto_merge": {
            "type": "boolean"
          },
          "allow_merge_commit": {
            "type": "boolean"
          },
          "allow_rebase_merge": {
            "type": "boolean"
          },
          "allow_squash_merge": {
            "type": "boolean"
          },
          "delete_branch_on_merge": {
            "type": "boolean"
          },
          "dismiss_stale_approvals": {
            "type": "boolean"
          },
          "require_codeowners": {
            "type": "boolean"
          },
          "require_exact_sha_approval": {
            "type": "boolean"
          },
          "require_jeryu_merge_passport": {
            "type": "boolean"
          },
          "require_linear_history": {
            "type": "boolean"
          },
          "required_approvals": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "NotificationSettings": {
        "type": "object",
        "required": [
          "watch_default",
          "notify_on_ci_failure",
          "notify_on_agent_completion",
          "notify_on_release"
        ],
        "properties": {
          "notify_on_agent_completion": {
            "type": "boolean"
          },
          "notify_on_ci_failure": {
            "type": "boolean"
          },
          "notify_on_release": {
            "type": "boolean"
          },
          "watch_default": {
            "type": "string"
          }
        }
      },
...
      "RepoDetailResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RepositorySummary"
          }
        ]
      },
...
      "RepositoryFacets": {
        "type": "object",
        "required": [
          "hosts",
          "owners",
          "families",
          "languages"
        ],
        "properties": {
          "families": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hosts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "owners": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "RepositoryId": {
        "type": "object",
        "description": "Opaque, stable repository identifier used on the API surface.\n\nPer §35.1.2 the canonical key in `/api/v1/repos/{repo_id}` is the opaque\n`id` (a UUID-shaped string persisted in `web_repositories.id`). The\nhuman-readable triple (`host`, `owner`, `name`) is preserved for display\nin the SPA so we can keep pretty URLs while the backend uses the stable id.",
        "required": [
          "id",
          "host",
          "owner",
          "name"
        ],
        "properties": {
          "host": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "description": "Opaque stable identifier (UUID-shaped). Canonical in API paths."
          },
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          }
        }
      },
      "RepositoryListResponse": {
        "type": "object",
        "required": [
          "generated_at",
          "total",
          "repositories",
          "facets"
        ],
        "properties": {
          "facets": {
            "$ref": "#/components/schemas/RepositoryFacets"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time"
          },
          "repositories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RepositorySummary"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "RepositorySettings": {
        "type": "object",
        "required": [
          "repo",
          "general",
          "features",
          "merge",
          "branch_protection",
          "ci",
          "agents",
          "access",
          "security",
          "notifications",
          "retention"
        ],
        "properties": {
          "access": {
            "$ref": "#/components/schemas/AccessSettings"
          },
          "agents": {
            "$ref": "#/components/schemas/AgentSettings"
          },
          "branch_protection": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchProtectionRule"
            }
          },
          "ci": {
            "$ref": "#/components/schemas/CiSettings"
          },
          "features": {
            "$ref": "#/components/schemas/FeatureSettings"
          },
          "general": {
            "$ref": "#/components/schemas/GeneralSettings"
          },
          "merge": {
            "$ref": "#/components/schemas/MergeSettings"
          },
          "notifications": {
            "$ref": "#/components/schemas/NotificationSettings"
          },
          "repo": {
            "$ref": "#/components/schemas/RepositoryId"
          },
          "retention": {
            "$ref": "#/components/schemas/RetentionSettings"
          },
          "security": {
            "$ref": "#/components/schemas/SecuritySettings"
          }
        }
      },
      "RepositorySummary": {
        "type": "object",
        "required": [
          "id",
          "entity",
          "visibility",
          "default_branch",
          "topics",
          "health",
          "open_merge_requests",
          "failing_checks",
          "running_jobs",
          "active_agents",
          "blocked_agents",
          "updated_at",
          "available_actions"
        ],
        "properties": {
          "active_agents": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "available_actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionRefSurrogate"
            }
          },
          "blocked_agents": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "clone_http_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "clone_ssh_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "default_branch": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "entity": {
            "$ref": "#/components/schemas/EntityRefSurrogate"
          },
          "failing_checks": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "family": {
            "type": [
              "string",
              "null"
            ]
          },
          "health": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/RepositoryId"
          },
          "language": {
            "type": [
              "string",
              "null"
            ]
          },
          "open_merge_requests": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "running_jobs": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "topics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "visibility": {
            "$ref": "#/components/schemas/RepositoryVisibility"
          }
        }
      },
      "RepositoryVisibility": {
        "type": "string",
        "enum": [
          "public",
          "internal",
          "private"
        ]
      },
...
      "RetentionSettings": {
        "type": "object",
        "required": [
          "audit_days",
          "evidence_days",
          "workflow_run_days",
          "log_days"
        ],
        "properties": {
          "audit_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "evidence_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "log_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "workflow_run_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "SecuritySettings": {
        "type": "object",
        "required": [
          "secret_scanning",
          "dependency_scanning",
          "license_policy_enabled",
          "agent_sandbox_required"
        ],
        "properties": {
          "agent_sandbox_required": {
            "type": "boolean"
          },
          "dependency_scanning": {
            "type": "boolean"
          },
          "license_policy_enabled": {
            "type": "boolean"
          },
          "secret_scanning": {
            "type": "boolean"
          }
        }
      },
...
      "SettingsPatch": {
        "type": "object",
        "description": "Settings patch wire-format. RFC 7396 merge semantics — only fields that\nappear in the JSON body are updated; absent fields are left unchanged.\nMirrors the host adapter's `HostRepositorySettingsPatch` for the subset\nof fields the BFF currently supports.",
        "properties": {
          "archived": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "default_branch": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "homepage": {
            "type": [
              "string",
              "null"
            ]
          },
          "visibility": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RepositoryVisibility"
              }
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jeryu_session"
      }
    }
  },
...
}